requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.9.0
argcomplete==3.6.3
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
email-validator==1.3.1
exceptiongroup==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
google-re2==1.1.20251105
googleapis-common-protos==1.72.0
graphviz==0.21
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python-ng==2.2.3
lazy-object-proxy==1.12.0
limits==5.6.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
more-itertools==10.8.0
msgspec==0.20.0
multidict==6.7.0
numpy==2.2.6
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==2.1.2
pillow==12.0.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pycparser==2.23
pydantic==1.10.26
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.30.0
setproctitle==1.3.7
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.6.2
Werkzeug==2.2.3
wrapt==2.0.1
WTForms==3.0.1
yarl==1.22.0
zipp==3.23.0
